Initial Commit, Fork from RADMBo project
Based on the progress made towards the RADMBo project, recycling all the
basic schematics for all the subsystems of the controller.
Revision one of the ZMBo schematic is very close to completion. What is
left is to build the configuration schematic, this where all the
different subsystems are set up from their "channel"  schematic files.
The configutation schematic is so that the ZMBo, ZMBo Lite, and ZMBo Pro
can all utilize the same subsystem channel schematics.